susciption controllers ok

+++ backend/gym/src/main/java/c1220ftjavareact/gym/domain/dto/SubscriptionDTO.java
@@ -11,8 +11,8 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 public class SubscriptionDTO {
-    private int idClient;
-    private int idClass;
+    private Long idClient;
+    private Long idClass;
     private State state;
     private LocalDateTime subscriptionDay;
 }
